requirements.txt
discord.py
requests
re
json